Add support for multiple date ranges
This commit adds support for multiple date ranges to the getCommitCount function. The user can now enter an array of date ranges, and the function will return the commit count for each range. The promptUserInput function has also been updated to prompt the user to enter the number of date ranges they want to enter and the start and end dates for each range.